+++ packages/rsc/src/browser.ts
@@ -21,8 +21,3 @@ export function initialize(): void {
     },
   });
 }
-
-import * as rscHtmlStreamClient from "rsc-html-stream/client";
-
-export const getRscStreamFromHtml = (): ReadableStream<Uint8Array> =>
-  rscHtmlStreamClient.rscStream;
