+++ packages/rsc/src/rsc-html-stream/browser.ts
@@ -1,0 +1,4 @@
+import * as rscHtmlStreamClient from "rsc-html-stream/client";
+
+export const getRscStreamFromHtml = (): ReadableStream<Uint8Array> =>
+  rscHtmlStreamClient.rscStream;
