+++ packages/rsc/src/ssr.tsx
@@ -76,11 +76,3 @@ export async function importRsc<T>(): Promise<T> {
 export function getAssetsManifest(): AssetsManifest {
   return (assetsManifest as any).default;
 }
-
-import * as rscHtmlStreamServer from "rsc-html-stream/server";
-
-export const injectRscStreamToHtml = (
-  stream: ReadableStream<Uint8Array>,
-  options?: { nonce?: string },
-): TransformStream<Uint8Array, Uint8Array> =>
-  rscHtmlStreamServer.injectRSCPayload(stream, options);
